Combine checkbox and checkbox_with_label
